Fix format strings for Windows.
Windows doesn't like %zd, so cast (s)size_t to int. Also, some shorts were
incorrectly printed with %d instead of %hd.

+++ src/net_socket.c
@@ -456,7 +456,7 @@ static void handle_meta_write(int sock, short events, void *data) {
 
 	ssize_t outlen = send(c->socket, c->outbuf.data + c->outbuf.offset, c->outbuf.len - c->outbuf.offset, 0);
 	if(outlen <= 0) {
-		logger(LOG_ERR, "Onoes, outlen = %zd (%s)", outlen, strerror(errno));
+		logger(LOG_ERR, "Onoes, outlen = %d (%s)", (int)outlen, strerror(errno));
 		terminate_connection(c, c->status.active);
 		return;
 	}
